acquire server refactoring
